refactor: extract proxy module
